Initial split of create-page logic
